1.9
полное обновлнеие за исключением смены скинов со скином

+++ Assets/ConvertHelper.cs
@@ -27,7 +27,7 @@ public class ConvertHelper : MonoBehaviour
         }
     }
 
-    public string GetCurrencyIntoString(float valueToConvert)
+    public string GetCurrencyIntoString(double valueToConvert)
     {
         if (valueToConvert >= 1000000000000f) // trillion
         {
